adds custom religions, holy books, posters, amulets, and so on

+++ code/game/turfs/floor_attackby.dm
@@ -600,7 +600,7 @@
 			user << "<span class='danger'>You can't plough this type of terrain.</span>"
 			return
 
-	else if (istype(C, /obj/item/weapon/covers) && !isturf(src, /turf/floor/beach/water/deep/saltwater))
+	else if (istype(C, /obj/item/weapon/covers) && !istype(src, /turf/floor/beach/water/deep/saltwater))
 
 		var/your_dir = "NORTH"
 
@@ -625,12 +625,13 @@
 			if (WWinput(user, "This will start building a floor cover [your_dir] of you.", "Floor Cover Construction", "Continue", list("Continue", "Stop")) == "Continue")
 				visible_message("<span class='danger'>[user] starts constructing the floor cover.</span>", "<span class='danger'>You start constructing the floor cover.</span>")
 				if (do_after(user, covers_time, user.loc))
-					qdel(C)
-					new/obj/covers/repairedfloor(src, user)
-					visible_message("<span class='danger'>[user] finishes placing the floor cover.</span>")
-					if (ishuman(user))
-						var/mob/living/carbon/human/H = user
-						H.adaptStat("crafting", 3)
+					if (!istype(src, /turf/floor/beach/water/deep/saltwater))
+						qdel(C)
+						new/obj/covers/repairedfloor(src, user)
+						visible_message("<span class='danger'>[user] finishes placing the floor cover.</span>")
+						if (ishuman(user))
+							var/mob/living/carbon/human/H = user
+							H.adaptStat("crafting", 3)
 				return
